Add export endpoint
In the same time, refactor tests to separate perf & rest tests so that "export" is not run after perf tests.
New common file for test resources definitions
