+++ hawkular-inventory-parent/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/InventoryService.java
@@ -119,4 +119,6 @@ public interface InventoryService {
      * @return true is InventoryService is Up and Running
      */
     boolean isRunning();
+
+    Import buildExport();
 }
