+++ hawkular-inventory-parent/hawkular-inventory-service/src/test/java/org/hawkular/inventory/service/InventoryRestTest.java
@@ -23,8 +23,6 @@ import java.io.File;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -41,21 +39,16 @@ import org.hawkular.inventory.api.Import;
 import org.hawkular.inventory.api.ResourceNode;
 import org.hawkular.inventory.api.ResourceWithType;
 import org.hawkular.inventory.api.ResultSet;
-import org.hawkular.inventory.model.Metric;
-import org.hawkular.inventory.model.MetricUnit;
-import org.hawkular.inventory.model.Operation;
 import org.hawkular.inventory.model.Resource;
 import org.hawkular.inventory.model.ResourceType;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.logging.Logger;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.shrinkwrap.resolver.api.maven.Maven;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.MethodSorters;
 
@@ -66,144 +59,6 @@ import org.junit.runners.MethodSorters;
 @RunWith(Arquillian.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class InventoryRestTest {
-    private final Logger log = Logger.getLogger(InventoryRestTest.class);
-
-    private static final Metric METRIC1
-            = new Metric("memory1", "Memory", MetricUnit.BYTES, new HashMap<>());
-    private static final Metric METRIC2
-            = new Metric("gc1", "GC", MetricUnit.NONE, new HashMap<>());
-    private static final Metric METRIC3
-            = new Metric("memory2", "Memory", MetricUnit.BYTES, new HashMap<>());
-    private static final Metric METRIC4
-            = new Metric("gc2", "GC", MetricUnit.NONE, new HashMap<>());
-
-    private static final Resource EAP1 = new Resource("EAP-1", "EAP-1", "feed1", "EAP", null,
-            Arrays.asList(METRIC1, METRIC2), new HashMap<>());
-    private static final Resource EAP2 = new Resource("EAP-2", "EAP-2", "feed2", "EAP", null,
-            Arrays.asList(METRIC3, METRIC4), new HashMap<>());
-    private static final Resource CHILD1 = new Resource("child-1", "Child 1", "feedX", "FOO", "EAP-1",
-            new ArrayList<>(), new HashMap<>());
-    private static final Resource CHILD2 = new Resource("child-2", "Child 2", "feedX", "BAR", "EAP-1",
-            new ArrayList<>(), new HashMap<>());
-    private static final Resource CHILD3 = new Resource("child-3", "Child 3", "feedX", "FOO", "EAP-2",
-            new ArrayList<>(), new HashMap<>());
-    private static final Resource CHILD4 = new Resource("child-4", "Child 4", "feedX", "BAR", "EAP-2",
-            new ArrayList<>(), new HashMap<>());
-    private static final Map<String, Map<String, String>> RELOAD_PARAMETERS;
-    static {
-        RELOAD_PARAMETERS = new HashMap<>();
-        RELOAD_PARAMETERS.put("param1", new HashMap<>());
-        RELOAD_PARAMETERS.get("param1").put("type", "bool");
-        RELOAD_PARAMETERS.get("param1").put("description", "Description of param1 for Reload op");
-        RELOAD_PARAMETERS.put("param2", new HashMap<>());
-        RELOAD_PARAMETERS.get("param2").put("type", "bool");
-        RELOAD_PARAMETERS.get("param2").put("description", "Description of param2 for Reload op");
-    }
-    private static final Map<String, Map<String, String>> SHUTDOWN_PARAMETERS;
-    static {
-        SHUTDOWN_PARAMETERS = new HashMap<>();
-        SHUTDOWN_PARAMETERS.put("param1", new HashMap<>());
-        SHUTDOWN_PARAMETERS.get("param1").put("type", "bool");
-        SHUTDOWN_PARAMETERS.get("param1").put("description", "Description of param1 for Shutdown op");
-        SHUTDOWN_PARAMETERS.put("param2", new HashMap<>());
-        SHUTDOWN_PARAMETERS.get("param2").put("type", "bool");
-        SHUTDOWN_PARAMETERS.get("param2").put("description", "Description of param2 for Shutdown op");
-    }
-    private static final Collection<Operation> EAP_OPS = Arrays.asList(
-            new Operation("Reload", RELOAD_PARAMETERS),
-            new Operation("Shutdown", SHUTDOWN_PARAMETERS),
-            new Operation("Start", Collections.emptyMap()));
-    private static final ResourceType TYPE_EAP = new ResourceType("EAP", EAP_OPS, new HashMap<>());
-    private static final ResourceType TYPE_FOO = new ResourceType("FOO", Collections.emptyList(), new HashMap<>());
-    private static final ResourceType TYPE_BAR = new ResourceType("BAR", Collections.emptyList(), new HashMap<>());
-
-    private static final Import IMPORT = new Import(Arrays.asList(EAP1, EAP2, CHILD1, CHILD2, CHILD3, CHILD4),
-            Arrays.asList(TYPE_EAP, TYPE_FOO, TYPE_BAR));
-
-    public static Import createResourceTypes() {
-        Map<String, Map<String, String>> reload;
-        reload = new HashMap<>();
-        reload.put("param1", new HashMap<>());
-        reload.get("param1").put("type", "bool");
-        reload.get("param1").put("description", "Description of param1 for Reload op");
-        reload.put("param2", new HashMap<>());
-        reload.get("param2").put("type", "bool");
-        reload.get("param2").put("description", "Description of param2 for Reload op");
-
-        Map<String, Map<String, String>> shutdown;
-        shutdown = new HashMap<>();
-        shutdown.put("param1", new HashMap<>());
-        shutdown.get("param1").put("type", "bool");
-        shutdown.get("param1").put("description", "Description of param1 for Shutdown op");
-        shutdown.put("param2", new HashMap<>());
-        shutdown.get("param2").put("type", "bool");
-        shutdown.get("param2").put("description", "Description of param2 for Shutdown op");
-        Collection<Operation> eapOps = Arrays.asList(
-                new Operation("Reload", reload),
-                new Operation("Shutdown", shutdown),
-                new Operation("Start", Collections.emptyMap()));
-        ResourceType eapType = new ResourceType("EAP", eapOps, new HashMap<>());
-
-        Map<String, Map<String, String>> flush;
-        flush = new HashMap<>();
-        flush.put("param1", new HashMap<>());
-        flush.get("param1").put("type", "bool");
-        flush.get("param1").put("description", "Description of param1 for Flush op");
-        flush.put("param2", new HashMap<>());
-        flush.get("param2").put("type", "bool");
-        flush.get("param2").put("description", "Description of param2 for Flush op");
-        Collection<Operation> jdgOps = Arrays.asList(
-                new Operation("Flush", flush),
-                new Operation("Delete", Collections.emptyMap()));
-        ResourceType jdgType = new ResourceType("JDG", jdgOps, new HashMap<>());
-
-        List<ResourceType> resourceTypes = Arrays.asList(eapType, jdgType);
-
-        return new Import(null, resourceTypes);
-    }
-
-    public static Import createLargeInventory(int from, int to, int children, int metrics) {
-        List<Resource> resources = new ArrayList<>();
-        for (int i = from; i < to; i++) {
-            String typeId = (i % 2 == 0) ? "EAP" : "JDG";
-            String id = "Server-" + i;
-            String name = "Server " + typeId + " with Id " + id;
-            String feedX = "feedX";
-            List<Resource> childrenResource = new ArrayList<>();
-            List<String> childrenIds = new ArrayList<>();
-            for (int j = 0; j < children; j++) {
-                String childType = (j % 2 == 0) ? "FOO" : "BAR";
-                String childIdX = id + "-child-" + j;
-                String childNameX = "Child "+ j + " from " + id;
-                childrenIds.add(childIdX);
-                Resource childX = new Resource(childIdX, childNameX, feedX, childType, id,
-                        new ArrayList<>(), new HashMap<>());
-
-                childrenResource.add(childX);
-            }
-
-            List<Metric> metricsResource = new ArrayList<>();
-            for (int k = 0; k < metrics; k++) {
-                Metric metricX = new Metric("metric-" + k, "Metric " + k, MetricUnit.BYTES, new HashMap<>());
-                metricsResource.add(metricX);
-            }
-
-            Map<String, String> propsX = new HashMap<>();
-            propsX.put("description", "This is a description for " + id);
-            Resource serverX = new Resource(id,
-                    name,
-                    feedX,
-                    typeId,
-                    null,
-                    metricsResource,
-                    propsX);
-
-            resources.add(serverX);
-            resources.addAll(childrenResource);
-        }
-
-        return new Import(resources, null);
-    }
 
     @ArquillianResource
     private URL baseUrl;
@@ -247,7 +102,7 @@ public class InventoryRestTest {
         WebTarget target = client.target(baseUrl.toString()).path("import");
         Response response = target
                 .request(MediaType.APPLICATION_JSON_TYPE)
-                .post(Entity.entity(IMPORT, MediaType.APPLICATION_JSON_TYPE));
+                .post(Entity.entity(Resources.IMPORT, MediaType.APPLICATION_JSON_TYPE));
         assertEquals(200, response.getStatus());
     }
 
@@ -461,84 +316,6 @@ public class InventoryRestTest {
     }
 
     @Test
-    @Category(Performance.class)
-    public void test018_largeImport() {
-        Client client = ClientBuilder.newClient();
-        WebTarget target = client.target(baseUrl.toString()).path("import");
-        Response response = target
-                .request(MediaType.APPLICATION_JSON_TYPE)
-                .post(Entity.entity(createResourceTypes(), MediaType.APPLICATION_JSON_TYPE));
-        assertEquals(200, response.getStatus());
-
-        int maxIterations = 50;
-        int maxServersPerIteration = 100;
-        int childrenPerServer = 100;
-        int metricsPerServer = 20;
-
-        for (int i = 0; i < maxIterations; i++) {
-            int from = i * maxServersPerIteration;
-            int to = ((i + 1) * maxServersPerIteration) - 1;
-            client = ClientBuilder.newClient();
-            target = client.target(baseUrl.toString()).path("import");
-            response = target
-                    .request(MediaType.APPLICATION_JSON_TYPE)
-                    .post(Entity.entity(createLargeInventory(from, to, childrenPerServer, metricsPerServer), MediaType.APPLICATION_JSON_TYPE));
-            assertEquals(200, response.getStatus());
-            int mod = maxIterations > 500 ? 100 : 10;
-            if ( i % mod == 0) {
-                log.infof("Creating [%s] Servers", (i * maxServersPerIteration));
-            }
-        }
-        log.infof("Final - Created [%s] Servers", (maxIterations * maxServersPerIteration));
-
-        log.infof("Fetch top resources");
-        for (int i = 0; i < maxIterations; i++) {
-            client = ClientBuilder.newClient();
-            target = client.target(baseUrl.toString())
-                    .path("resources")
-                    .queryParam("root", true)
-                    .queryParam("maxResults", maxServersPerIteration)
-                    .queryParam("startOffset", i * maxServersPerIteration);
-            response = target
-                    .request(MediaType.APPLICATION_JSON_TYPE)
-                    .accept(MediaType.APPLICATION_JSON_TYPE)
-                    .get();
-            assertEquals(200, response.getStatus());
-            int mod = maxIterations > 500 ? 100 : 10;
-            if ( i % mod == 0) {
-                log.infof("Querying [%s] Servers", (i * maxServersPerIteration));
-            }
-        }
-
-        log.info("Querying all types");
-        client = ClientBuilder.newClient();
-        target = client.target(baseUrl.toString()).path("types");
-        response = target
-                .request(MediaType.APPLICATION_JSON_TYPE)
-                .accept(MediaType.APPLICATION_JSON_TYPE)
-                .get();
-        assertEquals(200, response.getStatus());
-        assertThat((List<ResourceType>) response.readEntity(ResultSet.class).getResults())
-                .extracting(ResourceType::getId)
-                .containsOnly("EAP", "FOO", "BAR", "JDG");
-    }
-
-    @Test
-    @Category(Performance.class)
-    public void test019_shouldGetResourceTypesAndNotImpactFromResources() {
-        Client client = ClientBuilder.newClient();
-        WebTarget target = client.target(baseUrl.toString()).path("types");
-        Response response = target
-                .request(MediaType.APPLICATION_JSON_TYPE)
-                .accept(MediaType.APPLICATION_JSON_TYPE)
-                .get();
-        assertEquals(200, response.getStatus());
-        assertThat((List<ResourceType>) response.readEntity(ResultSet.class).getResults())
-                .extracting(ResourceType::getId)
-                .containsOnly("EAP", "FOO", "BAR", "JDG");
-    }
-
-    @Test
     public void test020_shouldGetAllEAPsPerFeed() {
         Client client = ClientBuilder.newClient();
         WebTarget target = client.target(baseUrl.toString()).path("resources")
@@ -568,11 +345,17 @@ public class InventoryRestTest {
     }
 
     @Test
-    public void zzz_clean() {
-        // FIXME: proper way for "AfterClass" with arquillian given there's non-static stuff needed?
-        // Delete resources
+    public void test021_shouldGetExport() {
         Client client = ClientBuilder.newClient();
-        WebTarget target = client.target(baseUrl.toString());
-        target.path("type/JDG").request().delete().close();
+        WebTarget target = client.target(baseUrl.toString()).path("export");
+        Response response = target
+                .request(MediaType.APPLICATION_JSON)
+                .get();
+        assertThat(response.getStatus()).isEqualTo(200);
+        Import imp = response.readEntity(Import.class);
+        assertThat(imp).isNotNull();
+        assertThat(imp.getResources()).extracting(Resource::getId).containsOnly("EAP-1", "EAP-2", "child-1",
+                "child-2", "child-3", "child-4", "CC", "CP");
+        assertThat(imp.getTypes()).extracting(ResourceType::getId).containsOnly("EAP", "FOO", "BAR");
     }
 }
